Small corrections (if ID is missing)

+++ www/klei00/cv07/delete.php
@@ -2,6 +2,9 @@
 require 'db.php';
 
 $productToDelete = $_GET['id'];
+if(!$productToDelete){
+    die("ID of a product is missing!");
+}
 $goodsDB->delete('id', $productToDelete);
 
 header('Location: index.php?delete=true');
